requirements.txt
asgiref==3.8.1
astroid==3.2.4
dill==0.3.8
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.2
PyJWT==2.9.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-decouple==3.8
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
tomlkit==0.13.0
